ams/requirements.txt
# Core Web Framework
fastapi==0.68.1
uvicorn==0.15.0

# AI Functionality
google-generativeai==0.3.1
python-dotenv==0.19.0
beautifulsoup4==4.9.3
aiohttp==3.8.1
asyncio==3.4.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0

# Authentication
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5

# Link Management
qrcode==7.4.2
Pillow==10.2.0  # Required for QR code generation

# Security
cryptography==42.0.2

# Data Validation
pydantic==1.8.2

# Background Tasks
celery==5.3.6
redis==5.0.1

# Reporting & Analytics
pandas==2.2.0
matplotlib==3.8.2

# Compliance & Tracking
requests==2.26.0

# Development & Testing
pytest==7.4.4
httpx==0.26.0
